requirements_tn.txt
beautifulsoup4
requests
PySimpleGUI
Pillow
